lextutorbackend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
motor==3.7.1
python-dotenv==1.0.0
python-multipart==0.0.20
passlib==1.7.4
bcrypt==4.3.0
python-jose==3.5.0
email-validator==2.2.0
pymongo==4.15.0
